requirements.txt
asgiref==3.8.1
crispy-bootstrap5==0.6
Django==4.0.10
django-cors-headers==3.10.1
djangorestframework==3.13.1
django-crispy-forms==1.13.0
django-allauth==0.50.0
drf-spectacular==0.21.2
sqlparse==0.5.3
typing_extensions==4.14.0
whitenoise==5.3.0
